STABLE10-sgi_65-compilation-patches-20010131
(cherry picked from commit 1ce1e6ed4900575fc341ace6c48e69d8fb98130b)

+++ src/rx/rx.c
@@ -455,7 +455,7 @@ int rx_Init(u_int port)
 #else
 	struct sockaddr_in addr;
 	int addrlen = sizeof(addr);
-	if (getsockname(rx_socket, (struct sockaddr *) &addr, &addrlen)) {
+	if (getsockname((int)rx_socket, (struct sockaddr *) &addr, &addrlen)) {
 	    rx_Finalize();
 	    return -1;
 	}
